calculates the Wind-Chill Factor (WCF) based on the air temperature and wind velocity entered by the user

+++ chapter 2/G/d.c
@@ -2,14 +2,12 @@
 #include <math.h>
 
 int main() {
-	float t, v; 
+	double t, v; 
 	printf("enter temperature in fahrenheit: ");
-	scanf("%f", &t);
+	scanf("%lf", &t);
 	printf("enter wind velocity in miles per hour: ");
-	scanf("%f", &v);
-
-	float wcf = 35.74 + 0.6215 * t + (0.4275 * t - 35.75) * pow(v, 0.16);
-	printf("wind-chill factor: %.2f", wcf);
-
+	scanf("%lf", &v);
+	double wcf = 35.74 + 0.6215 * t + (0.4275 * t - 35.75) * pow(v, 0.16);
+	printf("wind-chill factor: %.2lf", wcf);
 	return 0;
 }
